refactor: split server runtime concerns
